+++ Platform/OpenCanopy/OpenCanopy.c
@@ -37,12 +37,6 @@ typedef struct {
 } GUI_DRAW_REQUEST;
 
 //
-// Variables to assign the picked volume automatically once menu times out
-//
-extern BOOT_PICKER_GUI_CONTEXT mGuiContext;
-extern GUI_VOLUME_PICKER mBootPicker;
-
-//
 // I/O contexts
 //
 STATIC GUI_OUTPUT_CONTEXT            *mOutputContext    = NULL;
@@ -1143,13 +1137,14 @@ GuiDrawLoop (
   CONST LIST_ENTRY    *AnimEntry;
   CONST GUI_ANIMATION *Animation;
   UINT64              LoopStartTsc;
+  UINT64              LastTsc;
+  UINT64              NewLastTsc;
 
   ASSERT (DrawContext != NULL);
 
   mNumValidDrawReqs = 0;
   HoldObject        = NULL;
 
-  GuiRedrawAndFlushScreen (DrawContext);
   //
   // Clear previous inputs.
   //
@@ -1160,7 +1155,7 @@ GuiDrawLoop (
   //
   // Main drawing loop, time and derieve sub-frequencies as required.
   //
-  LoopStartTsc = mStartTsc = AsmReadTsc ();
+  LastTsc = LoopStartTsc = mStartTsc = AsmReadTsc ();
   do {
     if (mPointerContext != NULL) {
       //
@@ -1252,17 +1247,37 @@ GuiDrawLoop (
     //
     GuiFlushScreen (DrawContext);
 
+    NewLastTsc = AsmReadTsc ();
+
+    if (DrawContext->GuiContext->AudioPlaybackTimeout >= 0
+      && DrawContext->GuiContext->PickerContext->PickerAudioAssist) {
+      DrawContext->GuiContext->AudioPlaybackTimeout -= (INT32) (DivU64x32 (
+        GetTimeInNanoSecond (NewLastTsc - LastTsc),
+        1000000
+        ));
+      if (DrawContext->GuiContext->AudioPlaybackTimeout <= 0) {
+        DrawContext->GuiContext->PickerContext->PlayAudioFile (
+          DrawContext->GuiContext->PickerContext,
+          OcVoiceOverAudioFileSelected,
+          FALSE
+          );
+        DrawContext->GuiContext->PickerContext->PlayAudioEntry (
+          DrawContext->GuiContext->PickerContext,
+          DrawContext->GuiContext->BootEntry
+          );
+      }
+    }
+
     //
     // Exit early if reach timer timeout and timer isn't disabled due to key event
     //
     if (TimeOutSeconds > 0
-      && GetTimeInNanoSecond (AsmReadTsc () - LoopStartTsc) >= TimeOutSeconds * 1000000000ULL) {
-      //
-      // FIXME: There should be view function or alike.
-      //
-      mGuiContext.BootEntry = mBootPicker.SelectedEntry->Context;
+      && GetTimeInNanoSecond (NewLastTsc - LoopStartTsc) >= TimeOutSeconds * 1000000000ULL) {
+      DrawContext->GuiContext->ReadyToBoot = TRUE;
       break;
     }
+
+    LastTsc = NewLastTsc;
   } while (!DrawContext->ExitLoop (DrawContext->GuiContext));
 }